+++ changelog.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] 22.03.2026
+
+### Added
+
+- `matrix()` function
+- `infos()` function
+- `decode()` function
+
+### Changed
+
+- SVG Grayscaling simplified
+- bumped `ìmage` crate to 0.25
+
 ## [0.5.0] 2025-12-02
 
 ### Changed